requirements.txt
requests
confluent-kafka
tinybird-cli
pyarrow
python-dotenv